Moved sequence_gen to be a Sprite method
